refactor: 연관관계, 트랜잭션, 디렉토리 변경
