+++ src/main/java/roomescape/member/model/LoginMember.java
@@ -1,4 +1,4 @@
-package roomescape.member;
+package roomescape.member.model;
 
 public class LoginMember {
     private Long id;
