+++ src/main/java/roomescape/member/repository/MemberRepository.java
@@ -1,10 +1,12 @@
-package roomescape.member;
+package roomescape.member.repository;
 
 import java.util.NoSuchElementException;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import roomescape.member.model.Member;
 
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
